Integers-to-Long
The official API says that all msgs ID can be greater than 32bits at
any moment.

+++ src/main/java/de/vivistra/telegrambot/model/message/ContactMessage.java
@@ -15,7 +15,7 @@ public class ContactMessage extends Message {
 	 * @param recipient
 	 * @param message
 	 */
-	public ContactMessage(Integer recipient, Contact contact) {
+	public ContactMessage(Long recipient, Contact contact) {
 		super(recipient, contact);
 	}
